added wiktionary webscraper
the first few blocks of the datasetGenerator jupyternotebook are a test that uses beutiful soup to scrape the html table. The last few blocks use pandas which is a allot easier but has a little less control. the csv generation is currently implmented using the pandas method folders are based on part of speech verb noun adjective.